+++ app/src/main/java/com/example/des/hp/AttractionArea/AttractionAreaDetailsView.java
@@ -8,7 +8,6 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
@@ -24,6 +23,8 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 import static com.example.des.hp.Database.DatabaseAccess.databaseAccess;
+
+import androidx.annotation.NonNull;
 
 public class AttractionAreaDetailsView extends BaseActivity
 {
@@ -47,8 +48,8 @@ public class AttractionAreaDetailsView extends BaseActivity
             layoutName = "activity_attractionarea_details_view";
             setContentView(R.layout.activity_attractionarea_details_view);
             
-            txtAttractionAreaDescription = (TextView) findViewById(R.id.txtAttractionAreaDescription);
-            grpMenuFile = (LinearLayout) findViewById(R.id.grpMenuFile);
+            txtAttractionAreaDescription = findViewById(R.id.txtAttractionAreaDescription);
+            grpMenuFile = findViewById(R.id.grpMenuFile);
             
             afterCreate();
             
@@ -78,41 +79,33 @@ public class AttractionAreaDetailsView extends BaseActivity
     
     //region OnClick Events
     @Override
-    public boolean onOptionsItemSelected(MenuItem item)
+    public boolean onOptionsItemSelected(@NonNull MenuItem item)
     {
-        boolean lv_return = false;
         try
         {
-            switch (item.getItemId())
-            {
-                case R.id.action_edit_attractionarea:
-                    showAttractionAreaEdit();
-                    return (true);
-                case R.id.action_delete_attractionarea:
-                    deleteAttractionArea();
-                    return (true);
-                case R.id.action_view_attractionarea:
-                    showAttractionAreaView();
-                    return (true);
-                case R.id.action_add_generalattraction:
-                    StartNewAddIntent(GeneralAttractionDetailsEdit.class);
-                    return (true);
-                default:
-                    lv_return = super.onOptionsItemSelected(item);
-            }
+            int id=item.getItemId();
+            if(id==R.id.action_edit_attractionarea)
+                showAttractionAreaEdit();
+            if(id==R.id.action_delete_attractionarea)
+                deleteAttractionArea();
+            if(id==R.id.action_view_attractionarea)
+                showAttractionAreaView();
+            if(id==R.id.action_add_generalattraction)
+                StartNewAddIntent();
+            return(true);
         }
         catch (Exception e)
         {
             ShowError("onOptionsItemSelected", e.getMessage());
+            return(false);
         }
-        return (lv_return);
     }
     
-    public void StartNewAddIntent(Class neededClass)
+    public void StartNewAddIntent()
     {
         try
         {
-            Intent intent = new Intent(getApplicationContext(), neededClass);
+            Intent intent = new Intent(getApplicationContext(), GeneralAttractionDetailsEdit.class);
             intent.putExtra("ACTION", "add");
             intent.putExtra("HOLIDAYID", holidayId);
             intent.putExtra("DAYID", 0);
@@ -135,7 +128,7 @@ public class AttractionAreaDetailsView extends BaseActivity
     public void showForm()
     {
         super.showForm();
-        try(DatabaseAccess da = databaseAccess();)
+        try(DatabaseAccess da = databaseAccess())
         {
             allowCellMove = true;
             
@@ -152,7 +145,7 @@ public class AttractionAreaDetailsView extends BaseActivity
             txtAttractionAreaDescription.setText(attractionAreaItem.attractionAreaDescription);
             
             subTitle = attractionAreaItem.attractionAreaDescription;
-            if (title.length() > 0)
+            if (!title.isEmpty())
             {
                 SetTitles(title, subTitle);
             } else
@@ -164,15 +157,10 @@ public class AttractionAreaDetailsView extends BaseActivity
             
             CreateRecyclerView(R.id.attractionAreaListView, scheduleAdapter);
             
-            scheduleAdapter.setOnItemClickListener(new ScheduleAdapter.OnItemClickListener()
-            {
-                @Override
-                public void onItemClick(View view, ScheduleItem obj)
+            scheduleAdapter.setOnItemClickListener((view, obj) -> {
+                if (obj.schedType == getResources().getInteger(R.integer.schedule_type_generalattraction))
                 {
-                    if (obj.schedType == getResources().getInteger(R.integer.schedule_type_generalattraction))
-                    {
-                        StartNewEditIntent(GeneralAttractionDetailsView.class, obj);
-                    }
+                    StartNewEditIntent(obj);
                 }
             });
             afterShow();
@@ -183,11 +171,11 @@ public class AttractionAreaDetailsView extends BaseActivity
         }
     }
     
-    public void StartNewEditIntent(Class neededClass, ScheduleItem obj)
+    public void StartNewEditIntent(ScheduleItem obj)
     {
         try
         {
-            Intent intent = new Intent(getApplicationContext(), neededClass);
+            Intent intent = new Intent(getApplicationContext(), GeneralAttractionDetailsView.class);
             intent.putExtra("ACTION", "view");
             intent.putExtra("HOLIDAYID", obj.holidayId);
             intent.putExtra("DAYID", obj.dayId);
@@ -226,7 +214,7 @@ public class AttractionAreaDetailsView extends BaseActivity
         try
         {
             attractionAreaItem.noteId = pNoteId;
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 da.updateAttractionAreaItem(attractionAreaItem);
             }
@@ -257,7 +245,7 @@ public class AttractionAreaDetailsView extends BaseActivity
         try
         {
             attractionAreaItem.infoId = pInfoId;
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 da.updateAttractionAreaItem(attractionAreaItem);
             }
@@ -307,7 +295,7 @@ public class AttractionAreaDetailsView extends BaseActivity
     {
         try
         {
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 if (!da.deleteAttractionAreaItem(attractionAreaItem))
                     return;
